Minimum Working SimulateBattle
This commit was made while returning to the project after a long period
of time, so a lot of small things were touched and changed. The most
potentially problematic are the changes to the sql and the database
resources.

The bare bones of SimulateBattle are complete, without any concern for
the specfics of class stats and what not. Additionally, the method is
tested and some comment formatting has been performed.